requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
async-generator==1.10
attrs==23.1.0
backports.functools-lru-cache==1.6.4
browser==66.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
distlib==0.3.6
exceptiongroup==1.1.1
Faker==18.10.1
filelock==3.12.0
future==0.18.3
h11==0.14.0
idna==3.4
iniconfig==2.0.0
jesth==0.0.8
outcome==1.2.0
packaging==23.1
pipenv==2023.6.2
platformdirs==3.5.1
pluggy==1.0.0
PySocks==1.7.1
pytest==7.3.2
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
selene==2.0.0rc1
selenium==4.9.1
shared==0.0.27
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.3.0
urllib3==2.0.3
virtualenv==20.23.0
virtualenv-clone==0.5.7
webdriver-manager==3.8.6
wsproto==1.2.0
